// src/types/design.ts
export interface Design {

  _id: string;

  title: string;

  slug: string;

  shortDescription: string;

  description: string;

  category: string;

  difficulty: string;

  thumbnail: string;

  githubUrl: string;

  liveUrl: string;

  technologies: string[];

  featured: boolean;

  status: string;

}